medical-assistant-app/requirements.txt
fastapi
uvicorn
streamlit
sqlmodel
python-multipart
pydantic
openai
httpx
python-dotenv
pytest